Added sliders to control the profiles start and end

+++ UnderDevelopment/GridCal/Gui/Main/GridCalMain.py
@@ -1241,7 +1241,7 @@ class MainGUI(QMainWindow):
             dialogue = ProfileInputGUI(parent=self,
                                        list_of_objects=objects, magnitude=magnitude,
                                        AlsoReactivePower=False)
-            dialogue.resize(1.61 * 600.0, 600.0)  # golden ratio
+            dialogue.resize(int(1.61 * 600.0), 600)  # golden ratio
             dialogue.exec()  # exec leaves the parent on hold
 
             if dialogue.time is not None:
